♻️ refactor truffle-networks
