+++ scripts/operations/test.ts
@@ -1,6 +1,6 @@
 import { Ganache } from '@hoprnet/hopr-testing'
 import { bash } from './utils'
-import networks from '../../truffle-networks.json'
+import networks from '../../truffle-networks'
 
 export default async () => {
   const ganache = new Ganache({
