--- truffle-networks.json
@@ -1,19 +1,0 @@
-{
-  "development": {
-    "host": "127.0.0.1",
-    "port": 9545,
-    "network_id": "*"
-  },
-  "test": {
-    "host": "127.0.0.1",
-    "port": 9545,
-    "network_id": "*"
-  },
-  "soliditycoverage": {
-    "host": "127.0.0.1",
-    "port": 8555,
-    "network_id": "*"
-  },
-  "rinkeby": { "gas": 6500000, "gasPrice": 1000000000, "network_id": "4" },
-  "kovan": { "gas": 6500000, "gasPrice": 1000000000, "network_id": "42" }
-}
